aquarium: api: move bootstrap to nodes endpoint
