+++ src/aquarium.py
@@ -32,7 +32,6 @@ from gravel.controllers.resources.storage import Storage
 from gravel.controllers.services import Services
 
 from gravel.api import (
-    bootstrap,
     orch,
     status,
     services,
@@ -162,7 +161,6 @@ def aquarium_factory(
         await shutdown_method(aquarium_app, aquarium_api)
 
     aquarium_api.include_router(local.router)
-    aquarium_api.include_router(bootstrap.router)
     aquarium_api.include_router(orch.router)
     aquarium_api.include_router(status.router)
     aquarium_api.include_router(services.router)
